Refactor to use Error() method on function errors when converting to tfprotov5/6.FunctionError
